Renamed the Convertor class to Renderer.
Separated the zone information about each thread into its own structure,
keeping a list of zones that correspond to the threads. As a side effect,
this enables a list of QLine objects to be passed directly to the painter.

Started writing a colours dock widget - when completed, this will contain
the colours for the current pattern.